api_yamdb/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
pytz==2022.7
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
